finish

+++ Binomial.java
@@ -25,7 +25,7 @@ public class Binomial {
 
 	// Computes the Binomial function, efficiently
 	public static int binomial(int n, int k) {
-		long[][] memo = new long[n + 1][k + 1];
+		int[][] memo = new int[n + 1][k + 1];
 		for (int i = 0; i <= n; i++) {
 			for (int j = 0; j <= k; j++) {
 				memo[i][j] = -1;
